+++ apps/webgl/shaders/fragment.glsl
@@ -1,0 +1,6 @@
+precision mediump float;
+varying vec3 vColor;
+
+void main() {
+    gl_FragColor = vec4(vColor, 1.0);
+}
