+++ apps/webgl/src/lib.rs
@@ -1,4 +1,3 @@
-use std::f32::consts::PI;
 use wasm_bindgen::prelude::*;
 use web_sys::{
     HtmlCanvasElement, WebGlBuffer, WebGlProgram, WebGlRenderingContext, WebGlUniformLocation,
@@ -31,60 +30,13 @@ impl WebGLRenderer {
         let vert_shader = utils::compile_shader(
             &context,
             WebGlRenderingContext::VERTEX_SHADER,
-            r#"
-            attribute vec2 position;
-            attribute vec3 color;
-            varying vec3 vColor;
-            uniform float u_time;
-            uniform vec2 u_cursor;
-            uniform float u_scroll;
-
-            void main() {
-                vec2 pos = position;
-                float t = pos.y;
-                float y_offset = -0.5 + t;
-                float amplitude = 0.4 - t * 0.3;
-                float phase = t * 3.14159 * 1.5;
-                float frequency = 1.0 + t * 0.3;
-
-                pos.y = amplitude * sin((pos.x * frequency * 6.28318) + phase) + y_offset;
-
-                float baseFrequency = 6.0;
-                float baseAmplitude = 0.08;
-                float baseWave = sin(pos.x * baseFrequency + u_time * 0.5) * baseAmplitude;
-                
-                float scrollEffect = u_scroll * 0.2;
-                float scrollWave = sin(pos.x * 3.0 - u_time * 0.3) * scrollEffect;
-                
-                float distanceToCursor = distance(pos, u_cursor);
-                float cursorInfluence = smoothstep(0.4, 0.0, distanceToCursor);
-                
-                float modifiedFrequency = mix(baseFrequency, baseFrequency * 1.3, cursorInfluence);
-                float modifiedAmplitude = mix(baseAmplitude, baseAmplitude * 1.3, cursorInfluence);
-                float modifiedPhase = cursorInfluence * sin(u_time * 2.0) * 1.5;
-                
-                float modifiedWave = sin(pos.x * modifiedFrequency + u_time * 0.5 + modifiedPhase) * modifiedAmplitude;
-                
-                float finalWave = mix(baseWave, modifiedWave, cursorInfluence) + scrollWave;
-                
-                pos.y += finalWave;
-                gl_Position = vec4(pos.x, pos.y, 0.0, 1.0);
-                
-                vColor = mix(color, vec3(1.0, 0.5, 0.2), cursorInfluence * 0.5);
-            }
-            "#,
+            include_str!("../shaders/vertex.glsl"),
         )?;
 
         let frag_shader = utils::compile_shader(
             &context,
             WebGlRenderingContext::FRAGMENT_SHADER,
-            r#"
-            precision mediump float;
-            varying vec3 vColor;
-            void main() {
-                gl_FragColor = vec4(vColor, 1.0);
-            }
-            "#,
+            include_str!("../shaders/fragment.glsl"),
         )?;
 
         let program = utils::link_program(&context, &vert_shader, &frag_shader)?;
@@ -128,7 +80,7 @@ impl WebGLRenderer {
                 vertices.push(x);
                 vertices.push(t);
 
-                let color = interpolate_color(t);
+                let color = utils::interpolate_color(t);
                 colors.push(color.0);
                 colors.push(color.1);
                 colors.push(color.2);
@@ -218,15 +170,4 @@ impl WebGLRenderer {
     pub fn resize(&self, width: u32, height: u32) {
         self.context.viewport(0, 0, width as i32, height as i32);
     }
-}
-
-fn interpolate_color(t: f32) -> (f32, f32, f32) {
-    let orange = (1.0, 0.5, 0.0);
-    let red = (1.0, 0.0, 0.0);
-
-    (
-        orange.0 * (1.0 - t) + red.0 * t,
-        orange.1 * (1.0 - t) + red.1 * t,
-        orange.2 * (1.0 - t) + red.2 * t,
-    )
 }
